requirements.txt
# ===================================================================
# tpRigToolkit-tools-namemanager requirements file
# ===================================================================

lucidity
tpRigToolkit-core
tpDcc-libs-nameit
tpDcc-tools-nameit